Re #1945 (misc): Get rid of 64-bit compiler warnings for pjsua_app
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@5461 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip-apps/src/pjsua/pjsua_app_common.c
@@ -52,9 +52,9 @@ int my_atoi(const char *cs)
 	return 0 - (int)pj_strtoul(&s);
     } else if (cs[0] == '+') {
 	s.ptr++, s.slen--;
-	return pj_strtoul(&s);
+	return (int)pj_strtoul(&s);
     } else {
-	return pj_strtoul(&s);
+	return (int)pj_strtoul(&s);
     }
 }
